pyproject.toml
[tool.poetry]
name = "spothero"
version = "0.1.0"
description = ""
authors = ["Debra Sparr <dsparr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
djangorestframework = "^3.15.1"
pytz = "^2024.1"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"
pytest = "^8.1.1"
pytest-django = "^4.8.0"
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
